Change element type of array of arrays from Long to Int.

+++ benchmarks/time/src/main/scala/strawman/collection/immutable/ImmutableArrayPrimitiveBenchmark.scala
@@ -20,7 +20,7 @@ class ImmutableArrayPrimitiveBenchmark {
   var size: Int = _
 
   var xs: ImmutableArray[Int] = _
-  var xss: scala.Array[ImmutableArray[Long]] = _
+  var xss: scala.Array[ImmutableArray[Int]] = _
   var randomIndices: scala.Array[Int] = _
 
   @Setup(Level.Trial)
